v1_socket_spam/app/requirements.txt
aiohttp
python-socketio[asyncio_client]
transformers
torch
redis